MH-5529
Since we can no longer use "copy -vcodec", we need a general ffmpeg statement to handle our reencode.  The folks on ffmpeg irc vociferously said not to use sameq and recommeded qscale to specify quality compression.  The ffmpeg documentation on sameq is paltry and misleading so its understandable that one would interpret that sameq is acceptable ("Use same video quality as source (implies VBR)").

The irc flks reasons for not using sameq were:
-h264 input qscales won't match, for example, non-h264 output
-raw input
-nonlinear mpeg-2 quantizer scale
-basically anything that doesn't use the linear mpeg quantizer (wmv, etc.)

I tried some very basic tests using qscale 2 and 5.  It seems like qscale 2 retains a very high quality but there is some significant savings in file size for mpeg2.  With one mpeg4 sample it looked like the file roughly doubled in size and the bitrate increased roughly twofold.    Recommendations on standard qscale are needed!  The audio would continue to be copied to the new container. 


git-svn-id: http://opencast.jira.com/svn/MH/trunk@9610 e6edf6fb-f266-4316-afb4-e53d95876a76